@@ -1,7 +1,7 @@
 name: contacts_service_example
 description: Demonstrates how to use the contacts_service plugin.
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: ">=3.0.3 <4.0.0"
 dependencies:
   flutter:
     sdk: flutter
