@@ -1,12 +1,11 @@
 name: contacts_service
 description: A Flutter plugin to retrieve and manage contacts on Android and iOS devices.
-version: 0.6.3
+version: 0.6.4
 homepage: https://github.com/lukasgit/flutter_contacts
-maintainer: Lukas Dickie (@lukasgit)
 
 environment:
-  sdk: '>=2.12.0-0 <3.0.0'
-  flutter: '>=1.12.0'
+  sdk: ">=3.0.3 <4.0.0"
+
 
 dependencies:
   flutter:
